Move ExtensionService from engine.core to engine
