+++ src/TestEngine/agents/Program.cs
@@ -9,8 +9,6 @@ using System.IO;
 using System.Security;
 using TestCentric.Common;
 using TestCentric.Engine.Internal;
-using TestCentric.Engine.Services;
-using TestCentric.Engine.Helpers;
 using NUnit.Engine;
 
 namespace TestCentric.Engine.Agents
