+++ src/TestEngine/testcentric.engine.core.tests/Drivers/NUnitNetStandardDriverTests.cs
@@ -9,7 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Xml;
 using TestCentric.Tests.Assemblies;
-using TestCentric.Engine.Helpers;
+using TestCentric.Engine.Internal;
 using NUnit.Framework;
 using NUnit.Engine;
 using NUnit.Engine.Extensibility;
