+++ src/TestEngine/testcentric.engine.core/Internal/ExceptionHelper.cs
@@ -9,7 +9,7 @@ using System.Reflection;
 using System.Text;
 using NUnit.Engine;
 
-namespace TestCentric.Engine.Helpers
+namespace TestCentric.Engine.Internal
 {
     public static class ExceptionHelper
     {
