+++ src/TestEngine/testcentric.engine.core/Internal/TcpChannelUtils.cs
@@ -6,12 +6,10 @@
 #if !NETSTANDARD1_6 && !NETSTANDARD2_0
 using System;
 using System.Collections.Generic;
-using System.Runtime.Remoting;
 using System.Runtime.Remoting.Channels;
 using System.Runtime.Remoting.Channels.Tcp;
 using System.Runtime.Serialization.Formatters;
 using System.Threading;
-using TestCentric.Engine.Helpers;
 
 namespace TestCentric.Engine.Internal
 {
