+++ src/TestEngine/testcentric.engine.tests/Internal/ProcessUtilsTests.cs
@@ -6,7 +6,7 @@
 using NUnit.Framework;
 using System.Text;
 
-namespace TestCentric.Engine.Helpers
+namespace TestCentric.Engine.Internal
 {
     public static class ProcessUtilsTests
     {
