+++ src/TestEngine/testcentric.engine/Internal/AssemblyHelper.cs
@@ -8,7 +8,7 @@ using System;
 using System.IO;
 using System.Reflection;
 
-namespace TestCentric.Engine.Helpers
+namespace TestCentric.Engine.Internal
 {
     /// <summary>
     /// AssemblyHelper provides static methods for working
