+++ src/TestEngine/testcentric.engine/Internal/ProcessUtils.cs
@@ -5,7 +5,7 @@
 
 using System.Text;
 
-namespace TestCentric.Engine.Helpers
+namespace TestCentric.Engine.Internal
 {
     public static class ProcessUtils
     {
