+++ src/TestEngine/testcentric.engine/Runners/WorkItemTracker.cs
@@ -8,7 +8,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Xml;
 using NUnit.Engine;
-using TestCentric.Engine.Helpers;
+using TestCentric.Engine.Internal;
 
 namespace TestCentric.Engine.Runners
 {
