+++ src/TestEngine/testcentric.engine/Services/AggregatingTestRunner.cs
@@ -6,7 +6,7 @@
 using System;
 using System.Collections.Generic;
 using NUnit.Engine;
-using TestCentric.Engine.Helpers;
+using TestCentric.Engine.Internal;
 
 namespace TestCentric.Engine.Runners
 {
