+++ src/TestEngine/testcentric.engine/Services/ExtensionService.cs
@@ -12,7 +12,6 @@ using Mono.Cecil;
 using NUnit.Engine;
 using NUnit.Engine.Extensibility;
 using TestCentric.Engine.Extensibility;
-using TestCentric.Engine.Helpers;
 using TestCentric.Engine.Internal;
 
 namespace TestCentric.Engine.Services
