+++ src/TestEngine/testcentric.engine/Services/ResultWriters/NUnit3XmlResultWriter.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Xml;
 using System.IO;
 using NUnit.Engine.Extensibility;
-using TestCentric.Engine.Helpers;
+using TestCentric.Engine.Internal;
 
 namespace TestCentric.Engine.Services
 {
