+++ src/TestEngine/testcentric.engine/Services/RuntimeFrameworkService.cs
@@ -12,7 +12,6 @@ using System.Linq;
 using Microsoft.Win32;
 using NUnit.Engine;
 using TestCentric.Common;
-using TestCentric.Engine.Helpers;
 using TestCentric.Engine.Internal;
 
 namespace TestCentric.Engine.Services
