+++ src/TestEngine/testcentric.engine/Services/TestAgency.cs
@@ -9,7 +9,6 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Diagnostics;
 using TestCentric.Common;
-using TestCentric.Engine.Helpers;
 using TestCentric.Engine.Internal;
 using NUnit.Engine;
 using TestCentric.Engine.Communication.Transports.Remoting;
